fix DS-17, DS-18, DS-19; .md file support
The Slack Bot will now ask three questions and record the responses to the three questions when scraping a message. If a message has already been scraped, the bot will send a message notifying the user.

fix issues: DS-17, DS-18, DS-19

DS-17: The slack bot scraped all messages after it asked for a title, not just one.
Solution: implemented a variation of Izy's suggestion. The bot now checks if the timestamp of a parent thread has been scraped before asking the user to write a title, rather then using a timer to check, thus contributing to fixing DS-17.
In addition, if a message is not in direct response (in the same thread, right after the message sent by the bot) to the question the bot asked, the message will not be scraped, therefore also contributing to fixing DS-17

DS-18: update the slackbot to ask the user to upload a picture and then save the image link.
Solution: every time a reply message is sent to a parent message with a round_pushpin reaction, the bot checks whether it has asked for a title.  if it did, only then it will ask for a supplemental image. when the user submits a link to an image (a file will not work), it will record the image link and write it to the markdown file in .md format. if the user chooses not to upload an image, they will type "None" and the bot will record "No supplemental image was provided." This satisfies DS-18.

DS-19: update the slackbot to ask the user to add supplemental information and add it to the markdown file.
the same process used for DS-18 is applied here. after the bot asked and recorded the image link, it will ask the user to add supplemental text to the markdown file. if the user chooses not to add supplemental info, they will type "None" and the bot will not record any additional text. This satisfies DS-19.

Other changes:

1. updated the welcome message when a user joins the channel (or when "start" is typed) to describe how to use the bot.

2. stores the questions that the bot asks as variables in onboarding_tutorial.py

3. changed some imported libraries to remove redundancy and adjusted code accordingly

4. the bot will not listen to the event "pin_added" in my commits as I am only using message reactions.

5. updated code that displays the welcome message when "start" is typed to prevent error messages

6. added two functions.
the first checks whether a message has been scraped already.
the second writes text into the markdown file.

7. updated comments to explain updated code.